refactor: split up table module
